backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
pyyaml==6.0.2
python-multipart==0.0.9

# ML (install torch separately, see below)
# torch - install via: pip install torch --index-url https://download.pytorch.org/whl/cu121
transformers==4.49.0
peft>=0.14.0
accelerate==1.3.0
# bitsandbytes - for 4-bit quantization (Linux/Docker only)
# bitsandbytes==0.44.1

# RAG
chromadb>=0.5.0
sentence-transformers>=3.0.0
